scratch-buffers2: add periodic reporting of byte counters
The counter based stats that we had in the initial scratch-buffers2
implementation is only an indication whether something overused
scratch buffers causing a large pile of memory to be used and never freed.

This patch adds a metric that counts the number of bytes which is
updated every 5 seconds, this makes it easier to track how much bytes syslog-ng
uses.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>